+++ src/main/java/fll/xml/GoalGroup.java
@@ -73,10 +73,7 @@ public class GoalGroup extends GoalElement {
    */
   @Override
   public double evaluate(final TeamScore teamScore) {
-    if (!teamScore.scoreExists()) {
-      return Double.NaN;
-    } else if (teamScore.isNoShow()
-        || teamScore.isBye()) {
+    if (teamScore.isNoShow()) {
       return 0D;
     }